+++ admin-react/src/views/Notifications.js
@@ -46,7 +46,7 @@ function Notifications() {
       message: (
         <div>
           <div>
-            Welcome to <b>Light Bootstrap Dashboard React</b> - a beautiful
+            Welcome to <b>Admin Dashboard</b> - a beautiful
             freebie for every web developer.
           </div>
         </div>