+++ admin-react/src/views/Upgrade.js
@@ -21,10 +21,10 @@ function Upgrade() {
           <Col className="ml-auto mr-auto" md="8">
             <Card>
               <div className="header text-center">
-                <h4 className="title">Light Bootstrap Dashboard PRO React</h4>
+                <h4 className="title">Admin Dashboard PRO React</h4>
                 <p className="category">
                   Are you looking for more components? Please check our Premium
-                  Version of Light Bootstrap Dashboard React.
+                  Version of Admin Dashboard.
                 </p>
                 <br></br>
               </div>